requirements.txt
cvxopt==1.2.0
jupyter==1.0.0
matplotlib==3.3.4
networkx==2.2
numpy==1.19.2
pandas==1.1.5
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.4.1
statsmodels==0.12.2
tensorflow-gpu==2.2.0
tqdm==4.59.0
pytorch==1.9.0